Fix connection pooling, add Tedious spec
node-pool has a default maximum pool size of 1 (!), so fix that by
setting a max size of 100. This is equal to the default max pool size in
.NET, and sounds as good as any other number.

We were also not releasing connections when errors happened, which
coupled with the pool size of 1, meant that callers would hang
indefinitely whenever they attempted a SQL command after at least one
command has failed. This is now fixed.

Also add a non-pooling option to the Tedious adapter.